Car class modified
